Add no file message

+++ lib/xcmultilingual/writer.rb
@@ -6,6 +6,13 @@ module Xcmultilingual
     end
 
     def write()
+      puts "+ UPDATING FILES" if @verbose
+
+      if !File.exist?("#{@destination}")
+        puts "  There is no destination file."
+        return
+      end
+
       File.open("#{@destination}", "r+") do |file|
         file.write("That is the text.")
       end
